pyproject.toml
[project]
name = "benchmark-cursor-model"
version = "0.1.0"
description = "Cursor model performance benchmarking tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "numpy>=2.3.1",
  "opencv-python>=4.11.0.86",
  "pillow>=11.3.0",
  "psutil>=7.0.0",
  "pyautogui>=0.9.54",
  "pygetwindow>=0.0.9",
  "screeninfo>=0.8.1",
  "python-dotenv>=1.0.0",
  "python-xlib>=0.33",
]

[project.optional-dependencies]
test = ["pytest>=7.4.3", "pytest-xvfb>=3.0.0", "pytest-html>=4.1.1", "coverage>=7.0.0"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
